[cleanup] Extract elkjs related code in a dedicated hook
