+++ packages/diagrams/frontend/sirius-components-diagrams/src/renderer/layout/elk/useElkLayout.types.ts
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+import { Edge, Node } from '@xyflow/react';
+import { LayoutOptions } from 'elkjs/lib/elk-api';
+import { EdgeData, NodeData } from '../../DiagramRenderer.types';
+
+export interface UseElkLayoutValue {
+  elkLayout: (
+    nodes: Node<NodeData, string>[],
+    edges: Edge<EdgeData>[],
+    layoutOptions: LayoutOptions
+  ) => Promise<Node<NodeData, string>[]>;
+}
